fix(ingest): fix metadata for custom python packages (#9391)
