+++ metadata-ingestion/scripts/custom_package_codegen.py
@@ -73,6 +73,8 @@ __version__ = "{package_version}"
 """
     )
 
+    (src_path / "py.typed").write_text("")
+
     (package_path / "setup.py").write_text(
         f"""{autogen_header}
 from setuptools import setup
@@ -87,6 +89,11 @@ setup(
         "avro-gen3=={_avrogen_version}",
         "acryl-datahub",
     ],
+    package_data={{
+        "{python_package_name}": ["py.typed"],
+        "{python_package_name}.models": ["schema.avsc"],
+        "{python_package_name}.models.schemas": ["*.avsc"],
+    }},
     entry_points={{
         "datahub.custom_packages": [
             "models={python_package_name}.models.schema_classes",
